DjangoRestFramework/app/requirements.txt
aioredis==1.3.1
apiclient==1.0.4
asgiref==3.4.1
astroid==2.7.3
async-timeout==3.0.1
attrs==21.2.0
autobahn==21.3.1
Automat==20.2.0
backports.entry-points-selectable==1.1.0
Brlapi==0.8.2
cachetools==4.2.0
certifi==2021.5.30
cffi==1.14.6
channels==3.0.4
channels-redis==3.3.0
chardet==4.0.0
chrome-gnome-shell==0.0.0
constantly==15.1.0
cryptography==3.4.8
cupshelpers==1.0
daphne==3.0.2
dbus-python==1.2.18
debtags==2.1
distlib==0.3.2
distro==1.6.0
distro-info==1.0
Django==3.2.7
django-filter==21.1
django-webpack-loader==1.3.0
djangorestframework==3.12.4
filelock==3.0.12
gunicorn==20.1.0
hiredis==2.0.0
httplib2==0.18.1
hyperlink==21.0.0
idna==2.10
incremental==21.3.0
isort==5.9.3
lazy-object-proxy==1.6.0
louis==3.18.0
MarkupSafe==2.0.1
mccabe==0.6.1
msgpack==1.0.2
oauth2client==4.1.3
oauthlib==3.1.1
Pillow==8.3.2
pipenv==2021.5.29
platformdirs==2.3.0
protobuf==3.14.0
pyaes==1.6.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycairo==1.16.2
pycparser==2.20
pycups==2.0.1
pycurl==7.44.1
PyGObject==3.40.1
pylint==2.10.2
pyOpenSSL==20.0.1
pysmbc==1.0.23
python-apt==2.2.1
python-debian==0.1.39
pytz==2020.5
pyxdg==0.27
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.6
service-identity==21.1.0
six==1.16.0
sqlparse==0.4.1
supervisor==4.2.2
toml==0.10.2
txaio==21.2.1
typing-extensions==3.10.0.2
ufw==0.36
unattended-upgrades==0.1
uritemplate==3.0.1
urllib3==1.26.5
virtualenv==20.7.2
wrapt==1.12.1
xdg==5
zope.interface==5.4.0
